03_my_implementations/chapter3_release_patterns/02_model_load_pattern/model_loader/requirements.txt
google-cloud-storage>=2.18.0
click>=8.1.0
